added di support for presenter instance
