+++ app/src/main/java/com/example/otchallenge/di/PresenterModule.kt
@@ -1,0 +1,13 @@
+package com.example.otchallenge.di
+
+import com.example.otchallenge.model.BooksListPresenter
+import com.example.otchallenge.model.BooksListPresenterImpl
+import dagger.Binds
+import dagger.Module
+
+@Module
+abstract class PresenterModule {
+
+    @Binds
+    abstract fun bindBookListPresenter(presenter: BooksListPresenterImpl): BooksListPresenter
+}
